Reverted CSS/JS naming convention, increase font weight, narrower post pages, removed resume, added gallery
Increased the default font-weight from 200 to 300 to make it easier to
read content on the dark background.

The post pages are now narrower to make it easier to read content.

Removed the resume page, because I don’t really have a need to for it.

I’ve added a gallery page where I’ll be adding photos and artwork in
the future!

CSS and JS files are once again using ?v=XXX to force a cache refresh.
It’s easier to maintain the code this way, and at a low risk that
someone is using a browser so old that it won’t update.